9th

+++ ft_strncmp.c
@@ -1,15 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strncmp.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: gkhaishb <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/01/25 12:52:53 by gkhaishb          #+#    #+#             */
+/*   Updated: 2023/01/25 12:53:29 by gkhaishb         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-int ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+	size_t	i;
 
-    i = 0;
-    if (n == 0)
-        return (0);
-    while (s1[i] && s1[i] == s2[i] && i < n - 1)
-    {
-        i++;
-    }
-    return ((unsigned char)(s1[i]) - (unsigned char)(s2[i]));
+	i = 0;
+	if (n == 0)
+		return (0);
+	while (s1[i] && s1[i] == s2[i] && i < n - 1)
+	{
+		i++;
+	}
+	return ((unsigned char)(s1[i]) - (unsigned char)(s2[i]));
 }
